Update README.md
This is a 3D jump 'n' run game developed in C# with Unity 3D. It is still very slimmed down, there are only 6 levels and, one after the other, they become more difficult and challenging. It's best to get your own impression and try it out. 
I have also uploaded the associated properties, so if you want to modify the game, you can open the project in Unity 3D and then edit it accordingly and design and improve it according to your wishes and preferences.